+++ packages/international-phone-input/src/index.ts
@@ -1,2 +1,4 @@
 export { InternationalPhoneInput } from './Component.responsive';
 export type { InternationalPhoneInputProps } from './types';
+export { countriesData } from './data/country-data';
+export type { CountriesData } from './data/country-data';