+++ packages/international-phone-input/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @alfalab/core-components-international-phone-input
 
+## 2.7.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@17.20.2
+    -   input-autocomplete@12.7.3
+
+## 2.7.0
+
+### Minor Changes
+
+<sup><time>05.02.2025</time></sup>
+
+### [#1566](https://github.com/core-ds/core-components/pull/1566)
+
+-   Экспорт массива countriesData и его типа (для модификации по месту использования)
+
+## 2.6.0
+
+### Minor Changes
+
+<sup><time>04.02.2025</time></sup>
+
+### [#1562](https://github.com/core-ds/core-components/pull/1562)
+
+-   добавлен опциональный проп customCountriesList, который можно использовать, чтобы переопределить дефолтный массив country-data
+
 ## 2.5.10
 
 ### Patch Changes